+++ sui-move-runtime/src/handles.rs
@@ -629,9 +629,19 @@ mod tests {
         ObjectOut, Owner, TransactionEffects, TransactionEffectsV2,
     };
 
+    #[sui_move::move_struct(address = "0x2", module = "object", abilities = "copy, drop, store")]
+    struct ID {
+        bytes: Address,
+    }
+
+    #[sui_move::move_struct(address = "0x2", module = "object", abilities = "store")]
+    struct UID {
+        id: ID,
+    }
+
     #[sui_move::move_struct(address = "0x1", module = "demo", abilities = "key")]
     struct Demo {
-        id: sui_move::types::UID,
+        id: UID,
     }
 
     #[test]