+++ sui-move/examples/derive_struct.rs
@@ -1,8 +1,19 @@
 use sui_move::move_struct;
+use sui_move::prelude::Address;
+
+#[move_struct(address = "0x2", module = "object", abilities = "copy, drop, store")]
+pub struct ID {
+    pub bytes: Address,
+}
+
+#[move_struct(address = "0x2", module = "object", abilities = "store")]
+pub struct UID {
+    pub id: ID,
+}
 
 #[move_struct(address = "0x1", module = "vault", abilities = "key, store")]
 pub struct Vault {
-    pub id: sui_move::types::UID,
+    pub id: UID,
     pub value: u64,
 }